scaffold interface for event store layer
